+++ src/parsers/problem/PTAProblemParser.ts
@@ -23,7 +23,12 @@ export class PTAProblemParser extends Parser {
     task.setTimeLimit(parseInt(/(\d+)/.exec(timeLimitStr)[1], 10));
     task.setMemoryLimit(parseInt(/(\d+)/.exec(memoryLimitStr)[1], 10));
 
-    const blocks = container.querySelectorAll('.rendered-markdown > pre > code:not(.hljs)');
+    const blocks = [...container.querySelectorAll('.rendered-markdown > pre > code:not(.hljs)')].filter(
+      el =>
+        el.parentElement.previousElementSibling.tagName === 'H3' &&
+        el.parentElement.previousElementSibling.textContent.includes('样例'),
+    );
+
     for (let i = 0; i < blocks.length - 1; i += 2) {
       task.addTest(blocks[i].textContent, blocks[i + 1].textContent);
     }